bot/requirements.txt
python-telegram-bot==20.3
psycopg2-binary==2.9.9
requests==2.31.0